Move getContainerPorts to a common ports file
